Many performance improvements, better profiling abilities, refactoring

+++ lib/core/src/examples/sprite_system.rs
@@ -1,17 +1,16 @@
 #![allow(missing_docs)]
 
-use crate::display::object::{DisplayObjectOps, DisplayObject};
+use crate::display::object::DisplayObjectOps;
 use crate::display::symbol::geometry::sprite::Sprite;
 use crate::display::symbol::geometry::sprite::SpriteSystem;
 use crate::display::world::*;
 use crate::prelude::*;
 use crate::system::web::set_stdout;
 
-use basegl_system_web::get_performance;
 use nalgebra::Vector2;
 use nalgebra::Vector3;
 use wasm_bindgen::prelude::*;
-use web_sys::Performance;
+
 
 
 #[wasm_bindgen]
@@ -84,7 +83,7 @@ pub fn on_frame
 
     if !frozen {
         *time += 1;
-        let t = (*time as f32 / 50.0) as f32;
+        let t = *time as f32 / 50.0;
         let length = sprites.len() as f32;
         for (i, sprite) in sprites.iter_mut().enumerate() {
             let i = i as f32;
